requirements.txt
streamlit==1.28.0
openai==1.31.0
runwayml>=3.4.0
python-dotenv==1.0.0
requests==2.32.0
